requirements.txt
pandas==2.0.3
duckdb==1.0.0
great_expectations==0.18.19
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-cloud-storage

#pip install -r requirements.txt